requirements.txt
eo-learn-core
bunch
tensorflow
numpy
tqdm
scipy
scikit-image
